backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
PyJWT==2.6.0
psycopg2-binary==2.9.9
celery==5.3.0
redis==4.5.0
scikit-learn==1.2.0
spacy==3.5.0
PyPDF2==3.0.1
pdfplumber==0.7.1
textblob==0.17.1
gunicorn==20.1.0